refactored routing for teamSheet OG image
